Added grow method in field class
I've added a grow method in the field class. It takes light, food and
water  as arguments.

First it checks if the amount of crops is more than 0, if yes the each
of the crops grow-method is called with the water and light that we
already had.

Then it checks if there is any animals. if yes then it adds upp all the
food that is needed to grow the animals. Then it checks if there is any
additional food.
Then it starts the actual growing of the animals. It checks if the
amount of food is more than the amount of food that the animal needs.
Checks there is any additional food. if yes give more food to the
animal. Call the animal grow-method with the amount of food that animal
will have and the amount of water we have