Task 14
Added auto_grow function. It takes the field object and amount of days
as arguments and grows the field with random values on the light, water
and food on the chosen amount of days.

added manual_grow function wich takes a field object as an arguments and
grows the field for 1 day with user inputed values on food, water and
light

Added three functions, one which prints out a menu for the crops, one
for the menu of the animals and a main menu.

I've added a function that returns the chosen value of the user

and some more