added automatic grow method nad update display
I've added a method that gets called when the automatic_grow_button is
clicked. The method calls the chosen crops grow method 30 times with
random light and water value each time. When done it calls the
update_crop_view_status method

This method edits the three QLineEdits that whe have. They take the new
values that the automatic grow mehtod gave to the crop object and sets
them as text