added add_vehicle_through_command method
I've added a add method that can add vehicle objects through the command
line.

First it ask the user of what type of vehicle they want to add, either
cars or snowmobiles

then it asks what maker and model the vehicle have

then it ask for the price and checks if the price is an int with a value
more than 0

it then ask for what year the vehicle was made. It also checks if the
inputed year is an int between 1900 and 2015

if the user want a car it then chekcs for the amount of doors the car
shall have