Added a change_values button and a mehtod to call
The btn_change_vehicle_values is added and it is calling the
change_values method when clicked.

The method takes the text from teh QLineEdits and sets the values of the
chosen vehicle to the values the user entered. After that it calls the
create_vehicle_radio_button so the new values is printed out