Remade change_values method
I've remade the change_values method so it now uses the two methods
model.set_car_values and model.set_sm_values.
These methods returns true if the entered values are accepted and false
if they are not. This is then used to call the unaccepted_values_window
if it returns false so the user can know what values are accepted